feat: ✨ add FullscreenStreamData Overlay data aggregation to S3StreamDataService
- add new Endpoint to get the data
- extend S3StreamDataService to load more data
- add aggregation methods to Repositories

+++ src/main/java/tv/strohi/twitch/strohkoenigbot/splatoon3saver/s3api/model/inner/Stats.java
@@ -43,6 +43,6 @@ public class Stats {
 	public static class CurrentWeaponPowerOrder implements Serializable {
 		private String id;
 		private Double order;
-		private Integer weaponPower;
+		private Double weaponPower;
 	}
 }
